code refactor: add a constructor, switch int to double

+++ Tournament Simulation/Team.cpp
@@ -1,6 +1,6 @@
 #include "Team.h"
 
-Team::Team(const std::string& name, int rating)
+Team::Team(const std::string& name, double rating)
 {
 	this->name = name;
 	this->rating = rating;
